feat(format_path_list): always show full relative path of current file
