+++ src/components/editor.rs
@@ -11,11 +11,11 @@ use crate::{
     buffer::Buffer,
     components::component::Component,
     edit::{Action, ActionGroup, Edit, EditTransaction},
+    format_path_list::format_path_list,
     lsp::completion::PositionalEdit,
     position::Position,
     rectangle::Rectangle,
     selection::{CharIndex, Selection, SelectionMode, SelectionSet},
-    utils::format_path_list,
 };
 use crate::{
     app::{Dispatches, RequestParams, Scope},
